+++ FoodCircles/FCSTimelineData.h
@@ -11,14 +11,14 @@
 
 @interface FCSTimelineData : NSObject
 
-@property (assign, nonatomic) int type;
-@property (assign, nonatomic) int total;
-@property (assign, nonatomic) int dataType;
+@property (assign, nonatomic) NSInteger type;
+@property (assign, nonatomic) NSInteger total;
+@property (assign, nonatomic) NSInteger dataType;
 @property (strong, nonatomic) NSString *year;
 @property (strong, nonatomic) NSString *month;
 @property (strong, nonatomic) NSDate *date;
 @property (strong, nonatomic) NSString *code;
-@property (assign, nonatomic) int qtyFed;
+@property (assign, nonatomic) NSInteger qtyFed;
 @property (assign, nonatomic) NSInteger minumumDiners;
 @property (strong, nonatomic) NSString *restaurantName;
 @property (strong, nonatomic) NSString *offerName;
